requirements.txt
aiohttp~=3.8.3
pandas~=1.5.3
requests~=2.28.2
Pillow~=9.4.0
selenium~=4.7.2
gspread~=5.7.2
python-dotenv~=0.21.1
oauth2client~=4.1.3